Renames script Player Movement to Player
