--- Assets/Scripts/PlayerMovement.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b494a1a43dc8b46efa7ad7584ced8e68
-timeCreated: 1487546355
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
